requirements.txt
# Core Framework
gradio>=4.0.0
python-multipart>=0.0.6

# AI & NLP
transformers>=4.35.0
torch>=2.0.0
nltk>=3.8
textstat>=0.7.3
huggingface-hub>=0.17.0

# Document Generation - PDF
reportlab>=4.0.0
fpdf2>=2.7.0
weasyprint>=60.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Document Generation - Word
python-docx>=1.1.0
docx2pdf>=0.1.8

# Document Generation - Others
markdown>=3.5
jinja2>=3.1.2

# Data & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Tables & Formatting
tabulate>=0.9.0
beautifulsoup4>=4.12.0

# Utilities
pillow>=10.0.0
python-dateutil>=2.8.0
requests>=2.31.0

# LaTeX Support
pypandoc>=1.11

# Development & Utilities
python-dotenv>=1.0.0
